modularize transaction
