+++ dummydata/Transactions/__init__.py
@@ -1,0 +1,14 @@
+from Transactions.Transactions import Transaction
+from Transactions.SubscriptionTransaction import SubscriptionTransaction
+from Transactions.UberTransaction import UberTransaction
+from Transactions.AmazonTransaction import AmazonTransaction
+from Transactions.CoffeeTransaction import CoffeeTransaction
+from Transactions.SalaryTransaction import SalaryTransaction
+from Transactions.TransportationTransaction import TransportationTransaction
+from Transactions.UtilityTransaction import UtilityTransaction
+from Transactions.RentalTransaction import RentalTransaction
+from Transactions.BillPaymentTransaction import BillPaymentTransaction
+from Transactions.UtilityTransaction import UtilityTransaction
+from Transactions.NetflixTransaction import NetflixTransaction
+from Transactions.GroceryTransaction import GroceryTransaction
+from Transactions.ChipotleTransaction import ChipotleTransaction
