+++ dummydata/main.py
@@ -1,6 +1,7 @@
 import json
 from datetime import datetime, timedelta
 import random
+from Transactions import *
 
 example_account = {
     "account": {
@@ -121,343 +122,19 @@ example_account = {
   }
 
 
-class Transaction:
-    def __init__(self, transaction_id, account_id, date, amount, category, merchant, location, type):
-        self.transaction_id = transaction_id
-        self.account_id = account_id
-        self.date = date
-        self.amount = round(amount, 2)
-        self.category = category
-        self.merchant = merchant
-        self.location = location
-        self.type = type
-        self.status = "Completed"  # Default status
-        self.timestamp = datetime.now().strftime("%Y-%m-%dT%H:%M:%S")
-        self.payment_method = "Debit Card"  # Default payment method
-        self.object = {
-            "transaction_id": self.transaction_id,
-            "account_id": self.account_id,
-            "date": self.date,
-            "amount": self.amount,
-            "category": self.category,
-            "merchant": self.merchant,
-            "location": self.location,
-            "type": self.type
-        }
 
-    def create_transaction(self):
-        return {
-            "transaction_id": self.transaction_id,
-            "account_id": self.account_id,
-            "date": self.date,
-            "amount": self.amount,
-            "category": self.category,
-            "merchant": self.merchant,
-            "location": self.location,
-            "type": self.type,
-            "status": self.status,
-            "timestamp": self.timestamp,
-            "payment_method": self.payment_method
-        }
 
-class ChipotleTransaction(Transaction):
-    def __init__(self, transaction_id, account_id, date, amount, location="Manhattan, New York, NY"):
-        super().__init__(
-            transaction_id=transaction_id,
-            account_id=account_id,
-            date=date,
-            amount=-12.99,  # Standard Chipotle meal price
-            category="Dining",
-            merchant="Chipotle Mexican Grill",
-            location=location,
-            type="Debit"
-        )
-        self.object = {
-            "transaction_id": self.transaction_id,
-            "account_id": self.account_id,
-            "date": self.date,
-            "amount": self.amount,
-            "category": self.category,
-            "merchant": self.merchant,
-            "location": self.location,
-            "type": self.type
-        }
 
-class NetflixTransaction(Transaction):
-    def __init__(self, transaction_id, account_id, date):
-        super().__init__(
-            transaction_id=transaction_id,
-            account_id=account_id,
-            date=date,
-            amount=-12.99,
-            category="Entertainment",
-            merchant="Netflix",
-            location="Online",
-            type="Recurring Debit"
-        )
-        self.payment_method = "Subscription"
-        self.object = {
-            "transaction_id": self.transaction_id,
-            "account_id": self.account_id,
-            "date": self.date,
-            "amount": self.amount,
-            "category": self.category,
-            "merchant": self.merchant,
-            "location": self.location,
-            "type": self.type
-        }
 
-class RentalTransaction(Transaction):
-    def __init__(self, transaction_id, account_id, date, amount, merchant="Park Avenue Apartments"):
-        super().__init__(
-            transaction_id=transaction_id,
-            account_id=account_id,
-            date=date,
-            amount=-abs(amount),  # Ensure amount is negative
-            category="Rent",
-            merchant=merchant,
-            location="Manhattan, New York, NY",
-            type="ACH Transfer"
-        )
-        self.payment_method = "ACH"
-        self.object = {
-            "transaction_id": self.transaction_id,
-            "account_id": self.account_id,
-            "date": self.date,
-            "amount": self.amount,
-            "category": self.category,
-            "merchant": self.merchant,
-            "location": self.location,
-            "type": self.type
-        }
 
-class BillPaymentTransaction(Transaction):
-    def __init__(self, transaction_id, account_id, date, amount):
-        super().__init__(
-            transaction_id=transaction_id,
-            account_id=account_id,
-            date=date,
-            amount=-abs(amount),
-            category="Bill Payment",
-            merchant="Bill Payment",
-            location="New York, NY",
-            type="Bill Payment"
-        )
 
-class FreelanceTransaction(Transaction):
-    def __init__(self, transaction_id, account_id, date, amount):
-        super().__init__(
-            transaction_id=transaction_id,
-            account_id=account_id,
-            date=date,
-            amount=abs(amount),  # Ensure amount is positive
-            category="Income",
-            merchant="Freelance Work",
-            location="Manhattan, New York, NY",
-            type="Credit"
-        )
-        self.status = "Pending"
-        self.payment_method = "Bank Transfer"
-        self.object = {
-            "transaction_id": self.transaction_id,
-            "account_id": self.account_id,
-            "date": self.date,
-            "amount": self.amount,
-            "category": self.category,
-            "merchant": self.merchant,
-            "location": self.location,
-            "type": self.type
-        }
 
-class GroceryTransaction(Transaction):
-    def __init__(self, transaction_id, account_id, date, amount, merchant="Trader Joe's"):
-        super().__init__(
-            transaction_id=transaction_id,
-            account_id=account_id,
-            date=date,
-            amount=-abs(amount),  # Ensure amount is negative
-            category="Groceries",
-            merchant=merchant,
-            location="Manhattan, New York, NY",
-            type="Debit"
-        )
-        self.object = {
-            "transaction_id": self.transaction_id,
-            "account_id": self.account_id,
-            "date": self.date,
-            "amount": self.amount,
-            "category": self.category,
-            "merchant": self.merchant,
-            "location": self.location,
-            "type": self.type
-        }
 
-class UtilityTransaction(Transaction):
-    def __init__(self, transaction_id, account_id, date, amount, merchant="ConEdison", utility_type="Electricity"):
-        super().__init__(
-            transaction_id=transaction_id,
-            account_id=account_id,
-            date=date,
-            amount=-abs(amount),  # Ensure amount is negative
-            category="Utilities",
-            merchant=merchant,
-            location="New York, NY",
-            type="Bill Payment"
-        )
-        self.payment_method = "ACH"
-        self.utility_type = utility_type
-        self.object = {
-            "transaction_id": self.transaction_id,
-            "account_id": self.account_id,
-            "date": self.date,
-            "amount": self.amount,
-            "category": self.category,
-            "merchant": self.merchant,
-            "location": self.location,
-            "type": self.type
-        }
-class TransportationTransaction(Transaction):
-    def __init__(self, transaction_id, account_id, date, amount=127.00, merchant="MTA"):
-        super().__init__(
-            transaction_id=transaction_id,
-            account_id=account_id,
-            date=date,
-            amount=-abs(amount),
-            category="Transportation",
-            merchant=merchant,
-            location="New York, NY",
-            type="Recurring Debit"
-        ) 
-        self.payment_method = "Credit Card"
-        self.object = {
-            "transaction_id": self.transaction_id,
-            "account_id": self.account_id,
-            "date": self.date,
-            "amount": self.amount,
-            "category": self.category,
-            "merchant": self.merchant,
-            "location": self.location,
-            "type": self.type
-        }
 
-class SalaryTransaction(Transaction):
-    def __init__(self, transaction_id, account_id, date, amount, merchant="Tech Corp Inc"):
-        super().__init__(
-            transaction_id=transaction_id,
-            account_id=account_id,
-            date=date,
-            amount=abs(amount),  # Ensure amount is positive
-            category="Income",
-            merchant=merchant,
-            location="New York, NY",
-            type="Direct Deposit"
-        )
-        self.payment_method = "ACH"
-        self.status = "Completed"
-        self.object = {
-            "transaction_id": self.transaction_id,
-            "account_id": self.account_id,
-            "date": self.date,
-            "amount": self.amount,
-            "category": self.category,
-            "merchant": self.merchant,
-            "location": self.location,
-            "type": self.type
-        }
 
-class CoffeeTransaction(Transaction):
-    def __init__(self, transaction_id, account_id, date, amount=5.75, merchant="Starbucks"):
-        super().__init__(
-            transaction_id=transaction_id,
-            account_id=account_id,
-            date=date,
-            amount=-abs(amount),
-            category="Dining",
-            merchant=merchant,
-            location="Manhattan, New York, NY",
-            type="Debit"
-        )
-        self.object = {
-            "transaction_id": self.transaction_id,
-            "account_id": self.account_id,
-            "date": self.date,
-            "amount": self.amount,
-            "category": self.category,
-            "merchant": self.merchant,
-            "location": self.location,
-            "type": self.type
-        }
 
-class AmazonTransaction(Transaction):
-    def __init__(self, transaction_id, account_id, date, amount, item_category="Shopping"):
-        super().__init__(
-            transaction_id=transaction_id,
-            account_id=account_id,
-            date=date,
-            amount=-abs(amount),
-            category=item_category,
-            merchant="Amazon.com",
-            location="Online",
-            type="Debit"
-        )
-        self.payment_method = "Credit Card"
-        self.object = {
-            "transaction_id": self.transaction_id,
-            "account_id": self.account_id,
-            "date": self.date,
-            "amount": self.amount,
-            "category": self.category,
-            "merchant": self.merchant,
-            "location": self.location,
-            "type": self.type
-        }
 
-class UberTransaction(Transaction):
-    def __init__(self, transaction_id, account_id, date, amount, service_type="Ride"):
-        super().__init__(
-            transaction_id=transaction_id,
-            account_id=account_id,
-            date=date,
-            amount=-abs(amount),
-            category="Transportation" if service_type == "Ride" else "Food Delivery",
-            merchant=f"Uber{' Eats' if service_type == 'Food' else ''}",
-            location="New York, NY",
-            type="Debit"
-        )
-        self.object = {
-            "transaction_id": self.transaction_id,
-            "account_id": self.account_id,
-            "date": self.date,
-            "amount": self.amount,
-            "category": self.category,
-            "merchant": self.merchant,
-            "location": self.location,
-            "type": self.type
-        }
 
-class SubscriptionTransaction(Transaction):
-    def __init__(self, transaction_id, account_id, date, amount, merchant, category="Subscription"):
-        super().__init__(
-            transaction_id=transaction_id,
-            account_id=account_id,
-            date=date,
-            amount=-abs(amount),
-            category=category,
-            merchant=merchant,
-            location="Online",
-            type="Recurring Debit"
-        )
-        self.payment_method = "Subscription"
-        self.object = {
-            "account_id": self.account_id,
-            "account_name": self.account_name,
-            "account_type": self.account_type,
-            "account_number": self.account_number,
-            "balance": self.balance,
-            "owner_name": self.owner_name,
-            "bank_details": self.bank_details,
-            "transactions": self.transactions
-        }
