+++ server/main_test.go
@@ -391,3 +391,29 @@ func Test_main(t *testing.T) {
 		})
 	}
 }
+
+func Test_main(t *testing.T) {
+	tests := []struct {
+		name string // description of this test case
+	}{
+		// TODO: Add test cases.
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			main()
+		})
+	}
+}
+
+func Test_main(t *testing.T) {
+	tests := []struct {
+		name string // description of this test case
+	}{
+		// TODO: Add test cases.
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			main()
+		})
+	}
+}
